+++ src/themes/atom-light.ts
@@ -36,5 +36,7 @@ const palette = {
     hljsDeletionBg: '#ffeef0',
     hljsDeletionColor: '#e45649',
 };
+/** Background color for this theme. */
 export const bg = palette.bg;
+/** Complete CSS stylesheet for this theme. */
 export const css = buildThemeCss(palette);