+++ src/themes/coy.ts
@@ -35,5 +35,7 @@ const palette = {
     hljsDeletionBg: '#ffeef0',
     hljsDeletionColor: '#dd4a68',
 };
+/** Background color for this theme. */
 export const bg = palette.bg;
+/** Complete CSS stylesheet for this theme. */
 export const css = buildThemeCss(palette);