+++ src/themes/dracula.ts
@@ -35,5 +35,7 @@ const palette = {
     hljsDeletionBg: '#5b1818',
     hljsDeletionColor: '#ff5555',
 };
+/** Background color for this theme. */
 export const bg = palette.bg;
+/** Complete CSS stylesheet for this theme. */
 export const css = buildThemeCss(palette);